3d-models/backend/requirements.txt
fastapi
uvicorn
google-generativeai
chromadb
requests
python-dotenv
pydantic
groq
pillow
